fix: profile confirmation 기능 추가 및 post 시간 형식 추가

+++ static/js/confirmation.js
@@ -19,17 +19,20 @@ chkbx.forEach((checkbox) => {
 
 /// 각 예약 정보 삭제
 function deleteButton(btn, r_id, u_id) {
-    axios({
-        method: "DELETE",
-        url: "/profile/reservation",
-        data: { r_id: r_id, u_id: u_id },
-    })
-        .then(function (res) {
-            btn.parentElement.remove();
+    const areYouSure = confirm("삭제하시겠습니까?");
+    if (areYouSure) {
+        axios({
+            method: "DELETE",
+            url: "/profile/reservation",
+            data: { r_id: r_id, u_id: u_id },
         })
-        .catch((error) => {
-            console.error("Error deleting reservation:", error);
-        });
+            .then(function (res) {
+                btn.parentElement.remove();
+            })
+            .catch((error) => {
+                console.error("Error deleting reservation:", error);
+            });
+    }
 }
 
 // 체크 된 예약 정보 전부 삭제
@@ -39,23 +42,28 @@ function deleteAllCheckedBtn() {
             return checkbox.parentElement.parentElement;
         },
     );
-    for (let checkedEl of checkedReservations) {
-        const r_id = checkedEl.querySelector(".check-option").getAttribute("data-r-id");
-        const u_id = checkedEl.querySelector(".check-option").getAttribute("data-u-id");
-        axios({
-            method: "DELETE",
-            url: "/profile/reservation",
-            data: {
-                r_id: r_id,
-                u_id: u_id,
-            },
-        })
-            .then(function (res) {
-                checkedEl.remove();
-                checkAll.checked = false;
-            })
-            .catch((error) => {
-                console.error("Error deleting reservation:", error);
-            });
+    if (checkedReservations.length > 0) {
+        const areYouSure = confirm("삭제하시겠습니까?");
+        if (areYouSure) {
+            for (let checkedEl of checkedReservations) {
+                const r_id = checkedEl.querySelector(".check-option").getAttribute("data-r-id");
+                const u_id = checkedEl.querySelector(".check-option").getAttribute("data-u-id");
+                axios({
+                    method: "DELETE",
+                    url: "/profile/reservation",
+                    data: {
+                        r_id: r_id,
+                        u_id: u_id,
+                    },
+                })
+                    .then(function (res) {
+                        checkedEl.remove();
+                        checkAll.checked = false;
+                    })
+                    .catch((error) => {
+                        console.error("Error deleting reservation:", error);
+                    });
+            }
+        }
     }
 }
